refactor place format in plugins/base. make hanlder factory it's own class. consolidate report class and walk class.
