+++ picopt/plugins/tar.py
@@ -24,7 +24,6 @@ from typing import TYPE_CHECKING, Any
 import filetype
 from typing_extensions import override
 
-from picopt.formats import FileFormat
 from picopt.plugins.base import (
     ArchiveHandler,
     Detector,
@@ -33,6 +32,7 @@ from picopt.plugins.base import (
     Route,
     Tool,
 )
+from picopt.plugins.base.format import FileFormat
 from picopt.plugins.base.tool import StdLibTool
 
 if TYPE_CHECKING:
