+++ picopt/walk/handler_factory.py
@@ -31,7 +31,6 @@ from confuse.templates import AttrDict
 from treestamps import Grovestamps
 
 from picopt import plugins as registry
-from picopt.formats import FileFormat
 from picopt.path import PathInfo
 from picopt.plugins.base import (
     ArchiveHandler,
@@ -39,7 +38,9 @@ from picopt.plugins.base import (
     Handler,
     ImageHandler,
 )
-from picopt.walk.detect_format import DetectFormat
+from picopt.plugins.base.format import FileFormat
+from picopt.printer import Printer
+from picopt.walk.detect_format import detect_format
 
 
 def _is_pipeline_available(handler_cls: type[Handler], config: AttrDict) -> bool:
@@ -59,8 +60,13 @@ def _is_pipeline_available(handler_cls: type[Handler], config: AttrDict) -> bool
     return len(stages) == len(handler_cls.PIPELINE)
 
 
-class HandlerFactory(DetectFormat):
-    """Handler factory for walker."""
+class HandlerFactory:
+    """Handler factory for creating format-appropriate handlers."""
+
+    def __init__(self, config: AttrDict, printer: Printer) -> None:
+        """Initialize with config and printer."""
+        self._config: AttrDict = config
+        self._printer: Printer = printer
 
     def _lookup_route(
         self,
@@ -183,7 +189,9 @@ class HandlerFactory(DetectFormat):
     ) -> tuple[FileFormat | None, type[Handler] | None, Mapping[str, Any]]:
         handler_cls: type[Handler] | None = None
         try:
-            file_format, info = self.detect_format(path_info)
+            file_format, info = detect_format(
+                path_info, keep_metadata=self._config.keep_metadata
+            )
             handler_cls = self._pick_handler_class(
                 file_format,
                 convert=path_info.convert,
@@ -234,8 +242,8 @@ class HandlerFactory(DetectFormat):
             **kwargs,
         )
 
+    @staticmethod
     def create_repack_handler(
-        self,
         config: AttrDict,
         unpack_handler: ContainerHandler,
     ) -> ContainerHandler:
